Travel agent AI/requirements.txt
streamlit>=1.32.0
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
pandas>=2.2.0
aiohttp>=3.9.0
asyncio>=3.4.3
amadeus>=9.0.0
googlemaps>=4.10.0
wikipedia-api>=0.6.0 